+++ CHANGELOG.md
@@ -25,6 +25,7 @@
 * The PfxPass parameter on order objects is now obfuscated when serialized to disk. (#207)
 * Added 'PfxPassSecure' parameter to 'New-PACertificate', 'New-PAOrder', and 'Set-PAOrder' which takes precedence over 'PfxPass' if specified. (#207)
 * BUYPASS_PROD and BUYPASS_TEST are now recognized shortcuts for the the buypass.com CA environments when you use `Set-PAServer`.
+* Added tab completion for `Set-PAServer -DirectoryUrl`
 
 
 ## 3.18.1 (2020-11-12)